Move the new routes into the learn/lang url tree
